+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto/AA_PID_Tuning_TestDelete.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.auto;
 
 import com.disnodeteam.dogecv.CameraViewDisplay;
 import com.disnodeteam.dogecv.DogeCV;
@@ -46,7 +46,7 @@ import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 
 
-@Autonomous(name="AA_PID_Tuning_TestDelete", group="Codebusters")
+@Autonomous(name="AA_PID_Tuning_TestDelete", group="Auto Test")
 //@Disabled
 public class AA_PID_Tuning_TestDelete extends LinearOpMode {
     //Detector declaration
