+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto/A_ParkRightBridge.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.auto;
 
 import com.disnodeteam.dogecv.CameraViewDisplay;
 import com.disnodeteam.dogecv.DogeCV;
@@ -46,7 +46,7 @@ import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 
 
-@Autonomous(name="A_ParkRightBridge", group="Codebusters")
+@Autonomous(name="A_ParkRightBridge", group="Auto Test")
 //@Disabled
 public class A_ParkRightBridge extends LinearOpMode {
     //Detector declaration
