+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto/D_Two_Stone_DropRightWall.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.auto;
 
 import com.disnodeteam.dogecv.CameraViewDisplay;
 import com.disnodeteam.dogecv.DogeCV;
@@ -46,7 +46,7 @@ import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 
 
-@Autonomous(name="D_Two_Stone_DropRightWall", group="Codebusters")
+@Autonomous(name="D_Two_Stone_DropRightWall", group="Auto Test")
 //@Disabled
 public class D_Two_Stone_DropRightWall extends LinearOpMode {
     //Detector declaration
