+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto/Softwarebot_Test3_Autonomous.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.auto;
 
 import com.disnodeteam.dogecv.CameraViewDisplay;
 import com.disnodeteam.dogecv.DogeCV;
@@ -40,7 +40,7 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 
-@Autonomous(name="Autonomous Test3", group="Codebusters")
+@Autonomous(name="Autonomous Test3", group="Auto Test")
 @Disabled
 public class Softwarebot_Test3_Autonomous extends LinearOpMode {
